pyproject.toml
[tool.poetry]
name = "test-gemini-linguo"
version = "0.1.0"
description = ""
authors = ["piopy"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlalchemy = "^2.0.41"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
google-generativeai = "^0.8.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
